Moved code for sending SMS and Email notifications into clients/clientController.js
